Post class now interacts with db.

+++ bin/main.dart
@@ -4,30 +4,34 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:bliss/bliss.dart';
+import 'package:postgresql/postgresql.dart';
 
 part 'src/post.dart';
 part 'src/post_index.dart';
 part 'src/util.dart';
 
-void main() {
+Future main() async {
 
-  new Server(InternetAddress.ANY_IP_V6, 0)
+  // Initialise database connection.
+  pgdb = await connect(Platform.environment['DATABASE_URL']);
+
+  new Server(InternetAddress.ANY_IP_V6, 8080)
 
     // Edit posts (admin)
     ..addHandler('POST', '/post', createPost)
-    ..addHandler('PUT', '/post', updatePost)
+    ..addHandler('PUT', '/post/:id', updatePost)
     ..addHandler('DELETE', '/post/:id', deletePost)
 
     // Retrieve indexes and post data
-    ..addHandler('GET', '/post/:year', getIndex)
-    ..addHandler('GET', '/post/:year/:month', getIndex)
+    ..addHandler('GET', '/post/index/:year', getIndex)
+    ..addHandler('GET', '/post/index/:year/:month', getIndex)
     ..addHandler('GET', '/post/:id', getPost)
 
     // Edit and retrieve drafts (admin)
     ..addHandler('POST', '/draft', (Map data) => createPost(data, draft: true))
-    ..addHandler('PUT', '/draft', (Map data) => updatePost(data, draft: true))
+    ..addHandler('PUT', '/draft/:id', (Map data) => updatePost(data, draft: true))
     ..addHandler('DELETE', '/draft/:id', (Map data) => deletePost(data, draft: true))
-    ..addHandler('GET', '/draft', (Map data) => getIndex(data, draft: true))
+    ..addHandler('GET', '/draft/index', (Map data) => getIndex(data, draft: true))
     ..addHandler('GET', '/draft/:id', (Map data) => getPost(data, draft: true))
 
     ..start();
@@ -51,8 +55,12 @@ Future<Map> createPost(Map postData, {bool draft: false}) async {
 Future<Map> deletePost(Map postData, {bool draft: false}) async {
 
   Post post = await Post.fromExisting(postData['id'], draft: draft);
+
+  if (post == null)
+    return apiResponse(-2, 'Post does not exist.');
+
   if (!await post.delete())
-    return apiResponse(-2, 'Could not create new post.');
+    return apiResponse(-3, 'Could not delete post.');
   else
     return apiResponse(0);
 
@@ -68,8 +76,14 @@ Future<Map> getIndex(Map searchData, {bool draft: false}) async {
 }
 
 Future<Map> getPost(Map postData, {bool draft: false}) async {
-  Post post = await Post.fromExisting(postData['post_id'], draft: draft);
+
+  Post post = await Post.fromExisting(postData['id'], draft: draft);
+
+  if (post == null)
+    return apiResponse(-2, 'Post does not exist.');
+
   return post.toMap()..addAll(apiResponse(0));
+
 }
 
 Future<Map> updatePost(Map postData, {bool draft: false}) async {
@@ -79,6 +93,9 @@ Future<Map> updatePost(Map postData, {bool draft: false}) async {
 
   Post post = await Post.fromExisting(postData['id'], draft: draft);
 
+  if (post == null)
+    return apiResponse(-2, 'Post does not exist.');
+
   if (postData['title'] != null)
     post.title = postData['title'];
 
@@ -86,7 +103,7 @@ Future<Map> updatePost(Map postData, {bool draft: false}) async {
     post.body = postData['body'];
 
   if (!await post.save())
-    return apiResponse(-2, 'Could not create new post.');
+    return apiResponse(-3, 'Could not update post.');
   else
     return apiResponse(0);
